+++ utils/beam_profile.py
@@ -6,6 +6,10 @@ from matplotlib.ticker import (MultipleLocator, AutoMinorLocator)
 import toml
 from datetime import datetime
 import os
+import tkinter as tk
+from tkinter import filedialog
+import glob
+
 
 class data_structure:  
     def __init__(self):
@@ -265,3 +269,24 @@ class data_structure:
             f.write('\n'.join(comments))
         self.fig_gaus.savefig(export_directory + "/OUTPUT_WIDTHS.png", bbox_inches='tight')
         self.fig_profile.savefig(export_directory + "/OUTPUT_BEAM_PROFILE.png", bbox_inches='tight')
+
+
+
+def select(data_structure):
+        '''
+        Opens a window to select the files you want to load
+
+        ## Generates:
+        - self.directory: list with the directories of the selected files
+        - self.folder: folder where the files are located
+        - self.names: list of the file names
+        '''
+
+        # Open Window
+        root = tk.Tk()
+        root.attributes('-topmost', True)
+        root.withdraw()
+        st.session_state['profile_directory'] = filedialog.askdirectory(title='Select Directory', initialdir=st.session_state['profile_directory'], parent=root)
+        root.destroy()
+
+        data_structure.all_files = glob.glob(st.session_state['profile_directory'] + '/Data_*.txt') # Load all files having a specific name format within the working directory)
